requirements.txt
streamlit
opencv-python
numpy
scikit-learn
Pillow
pandas
matplotlib
